+++ webextension/js/options-data.js
@@ -25,21 +25,12 @@ export const options = {
 		"prefLevel": "basic",
 		"onlyDelegatedMode": true
 	},
-	"sending_websites_data_support": {
-		"title": "Accept sending deviantArt/FreshRss data (external mode only)",
-		"description": "Enabled if checked",
-		"type": "bool",
-		"value": true,
-		"prefLevel": "basic",
-		"onlyDelegatedMode": true
-	},
 	"check_enabled": {
 		"title": "Data automatic refresh enabled",
 		"description": "Enabled if checked",
 		"type": "bool",
 		"value": true,
-		"prefLevel": "basic",
-		"onlyNormalMode": true
+		"prefLevel": "basic"
 	},
 	"check_delay": {
 		"title": "Streams status delay",